agent2/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
google-genai>=0.2.2
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pydantic>=2.0.0
a2a-sdk[http-server]
